participant_test/requirements.txt
torch
torchvision
tqdm
pyyaml
opencv-python
numpy
matplotlib
